PP-642: Little suggestion
