+++ src/hooks/useShowHelp.tsx
@@ -1,15 +1,33 @@
 import { useCallback, useContext } from 'react'
 import { OverlayContext } from '../contexts/overlay'
 import { useNavigation } from '../hooks'
-import { showHelp } from '../overlays/showHelp'
+import { helpOverlays } from '../overlays/showHelp'
+import i18n from '../utils/i18n'
 
-export const useShowHelp = () => {
+export const useShowHelp = (id: 'paymentMethods' | 'currencies') => {
   const navigation = useNavigation()
   const [, updateOverlay] = useContext(OverlayContext)
 
-  const showHelpContextualized = useCallback((i: HelpType) => {
-    showHelp(updateOverlay, i, navigation)
-  }, [])
+  const showHelp = useCallback(() => {
+    const goToHelp = () => {
+      updateOverlay({
+        visible: false,
+      })
+      navigation.navigate('contact')
+    }
 
-  return showHelpContextualized
+    updateOverlay({
+      title: helpOverlays.title[id],
+      content: helpOverlays.content[id],
+      visible: true,
+      action2: {
+        callback: goToHelp,
+        label: i18n('help'),
+        icon: 'info',
+      },
+      level: 'INFO',
+    })
+  }, [id, navigation, updateOverlay])
+
+  return showHelp
 }
